+++ packages/components/add-button/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.31](https://github.com/Synerise/synerise-design/compare/@synerise/ds-add-button@0.1.30...@synerise/ds-add-button@0.1.31) (2020-04-29)
+
+**Note:** Version bump only for package @synerise/ds-add-button
+
+
+
+
+
 ## [0.1.30](https://github.com/Synerise/synerise-design/compare/@synerise/ds-add-button@0.1.29...@synerise/ds-add-button@0.1.30) (2020-04-27)
 
 **Note:** Version bump only for package @synerise/ds-add-button