+++ packages/components/dropdown/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.28](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.11.27...@synerise/ds-dropdown@0.11.28) (2020-04-29)
+
+**Note:** Version bump only for package @synerise/ds-dropdown
+
+
+
+
+
 ## [0.11.27](https://github.com/synerise/synerise-design/compare/@synerise/ds-dropdown@0.11.26...@synerise/ds-dropdown@0.11.27) (2020-04-27)
 
 **Note:** Version bump only for package @synerise/ds-dropdown