+++ packages/components/items-roll/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.3](https://github.com/Synerise/synerise-design/compare/@synerise/ds-items-roll@0.1.2...@synerise/ds-items-roll@0.1.3) (2020-04-29)
+
+
+### Bug Fixes
+
+* **items-roll:** add style fixes ([9a3dbef](https://github.com/Synerise/synerise-design/commit/9a3dbef33d2ef4537c3c4ca97428a8d14cd95a07))
+
+
+
+
+
 ## [0.1.2](https://github.com/Synerise/synerise-design/compare/@synerise/ds-items-roll@0.1.1...@synerise/ds-items-roll@0.1.2) (2020-04-28)
 
 **Note:** Version bump only for package @synerise/ds-items-roll